+++ Source/Mosa.Compiler.Framework/Stages/BaseTransformStage.cs
@@ -1,10 +1,7 @@
 ﻿// Copyright (c) MOSA Project. Licensed under the New BSD License.
 
-using System.Collections;
 using System.Collections.Generic;
-using System.Diagnostics;
 using Mosa.Compiler.Framework.Trace;
-using Mosa.Compiler.Framework.Transforms.BasicBlocks;
 
 namespace Mosa.Compiler.Framework.Stages;
