+++ Source/Mosa.Compiler.Framework/Stages/DeadBlockStage.cs
@@ -1,5 +1,7 @@
 ﻿// Copyright (c) MOSA Project. Licensed under the New BSD License.
 
+using Mosa.Compiler.Framework.Analysis;
+
 namespace Mosa.Compiler.Framework.Stages;
 
 /// <summary>
