--- Source/Mosa.Compiler.Framework/Stages/IRCleanupStage.cs
@@ -1,54 +1,0 @@
-﻿// Copyright (c) MOSA Project. Licensed under the New BSD License.
-
-using Mosa.Compiler.Framework.Analysis;
-
-namespace Mosa.Compiler.Framework.Stages;
-
-/// <summary>
-/// IR Cleanup Stage
-/// </summary>
-public class IRCleanupStage : BaseTransformStage
-{
-	public IRCleanupStage() : base()
-	{
-		AddTranforms(Transforms.BasicBlocks.BasicBlocksTransforms.List);
-
-		EnableBlockOptimizations = true;
-	}
-
-	protected override void Run()
-	{
-		RemoveNops();
-
-		base.Run();
-
-		OrderBlocks();
-	}
-
-	private void RemoveNops()
-	{
-		foreach (var block in BasicBlocks)
-		{
-			for (var node = block.AfterFirst; !node.IsBlockEndInstruction; node = node.Next)
-			{
-				if (node.IsEmpty || !node.IsNop)
-					continue;
-
-				node.Empty();
-			}
-		}
-	}
-
-	private void OrderBlocks()
-	{
-		var blockOrderAnalysis = new SimpleTraceBlockOrder();
-
-		blockOrderAnalysis.Analyze(BasicBlocks);
-
-		var newBlockOrder = blockOrderAnalysis.NewBlockOrder;
-
-		newBlockOrder = AddMissingBlocksIfRequired(newBlockOrder);
-
-		BasicBlocks.ReorderBlocks(newBlockOrder);
-	}
-}
