+++ Source/Mosa.Compiler.Framework/Transforms/BasicBlocks/SkipEmptyBlocks.cs
@@ -39,9 +39,9 @@ public class SkipEmptyBlocks : BaseBlockTransform
 			if (!block.IsEmptyBlockWithSingleJump())
 				continue;
 
-			var hasPhi = block.NextBlocks[0].HasPhiInstruction();
+			var hasPhi = isInSSAForm && block.NextBlocks[0].HasPhiInstruction();
 
-			if (hasPhi && isInSSAForm && (block.PreviousBlocks.Count != 1 || block.NextBlocks[0].PreviousBlocks.Count != 1))
+			if (hasPhi && (block.PreviousBlocks.Count != 1 || block.NextBlocks[0].PreviousBlocks.Count != 1))
 				continue;
 
 			trace?.Log($"Removed Block: {block} - Skipped to: {block.NextBlocks[0]}");
